Use the native Swift struct initializers rather than use CGGeometry functions.
In Objective-c, we used to use CGRectMake to get a CGRect struct because for   initializing a strut, it is necessary ( as in C if my memory is good ) to create first the structure, and then assigning variable.

With Swift, struct have constructors with parameters, so no need to use external functions.

Signed-off-by: Rebouh Aymen <aymenworks@gmail.com>