Xcode 7 and Swift 2.0 fix
-Replacing var by let when possible
-Using Swift native constructors instead  functions that return CGPoint
- 0 for options is not accepted. Nil also. With Swift, it becomre [].

PS : In Objective-c, we used to use CGPointMake to get a CGPoint struct because for  initializing a strut, it is necessary ( as in C if my memory is good ) to create first the structure, and then assigning variable.

With Swift, struct have constructors with parameters, so no need to use external functions.

Signed-off-by: Rebouh Aymen <aymenworks@gmail.com>